+++ includes/plugin/class-uninstaller.php
@@ -29,7 +29,7 @@ class Uninstaller {
 	 * @since 1.0.0
 	 */
 	public static function uninstall() {
-		Option::delete_all();
+		Option::site_delete_all();
 		User::delete_all_meta();
 		// Delete cache?