+++ includes/system/class-assets.php
@@ -33,7 +33,7 @@ class Assets {
 	 * @since 1.0.0
 	 */
 	public function prefetch() {
-		if ( Option::get( 'use_cdn' ) && DECALOG_CDN_AVAILABLE ) {
+		if ( Option::network_get( 'use_cdn' ) && DECALOG_CDN_AVAILABLE ) {
 			echo '<meta http-equiv="x-dns-prefetch-control" content="on">';
 			echo '<link rel="dns-prefetch" href="//cdn.jsdelivr.net" />';
 		}
@@ -56,7 +56,7 @@ class Assets {
 	 * @since  1.0.0
 	 */
 	public function register_style( $handle, $src, $file, $deps = [], $media = 'all' ) {
-		if ( Option::get( 'use_cdn' ) && DECALOG_CDN_AVAILABLE ) {
+		if ( Option::network_get( 'use_cdn' ) && DECALOG_CDN_AVAILABLE ) {
 			if ( DECALOG_ADMIN_URL === $src ) {
 				$file = 'https://cdn.jsdelivr.net/wp/' . DECALOG_SLUG . '/tags/' . DECALOG_VERSION . '/admin/' . $file;
 			} else {
@@ -83,16 +83,16 @@ class Assets {
 	 * @since  1.0.0
 	 */
 	public function register_script( $handle, $src, $file, $deps = [] ) {
-		if ( Option::get( 'use_cdn' ) && DECALOG_CDN_AVAILABLE ) {
+		if ( Option::network_get( 'use_cdn' ) && DECALOG_CDN_AVAILABLE ) {
 			if ( DECALOG_ADMIN_URL === $src ) {
 				$file = 'https://cdn.jsdelivr.net/wp/' . DECALOG_SLUG . '/tags/' . DECALOG_VERSION . '/admin/' . $file;
 			} else {
 				$file = 'https://cdn.jsdelivr.net/wp/' . DECALOG_SLUG . '/tags/' . DECALOG_VERSION . '/public/' . $file;
 			}
 			// phpcs:ignore
-			return wp_register_script( $handle, $file, $deps, null, Option::get( 'script_in_footer' ) );
+			return wp_register_script( $handle, $file, $deps, null, Option::network_get( 'script_in_footer' ) );
 		} else {
-			return wp_register_script( $handle, $src . $file, $deps, DECALOG_VERSION, Option::get( 'script_in_footer' ) );
+			return wp_register_script( $handle, $src . $file, $deps, DECALOG_VERSION, Option::network_get( 'script_in_footer' ) );
 		}
 	}