+++ CHANGELOG.md
@@ -3,14 +3,17 @@ All notable changes to **DecaLog** is documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.0] - 2019-09-11
 ### Added
-- Support for [htaccess Server-Info & Server-Status](https://wordpress.org/plugins/htaccess-server-info-server-status/) plugin.
+- Listener for WordPress multisite specific events.
+- Listener for [htaccess Server-Info & Server-Status](https://wordpress.org/plugins/htaccess-server-info-server-status/) plugin.
 - Links to support, site and GitHub repository in plugin list.
 - New class (Decalog\Logger) to use as a standard PSR-3 logger (for plugins and themes developers).
 - New shortcodes to query the plugin statistics.
 ### Removed
 - WordPress release now excludes GitHub .wordpress-org directory.
+### Fixed
+- A WordPress logger may record log in the wrong table in multisites instances.
 
 ## [1.2.1] - 2019-09-05
 ### Changed